Remove some code duplication in Way and Relation classes
Fixes https://github.com/MarcusWolschon/osmeditor4android/issues/1945
